# TimeAdministrator/source/services/TimeDeskService/src/LAG.Services.TimeDesk/Controllers/Models/TimeLogResponse.cs
namespace EPY.Services.LogTiempo.Controllers.Models
{
    public class TimeLogResponse
    {
        public string Id { get; set; }

        public string StartTime { get; set; }

        public string EndTime { get; set; }

        public string Type { get; set; }
    }
}